refactor: extract multiband EQ module
