+++ src/components/recorder/eq-response-graph.tsx
@@ -7,8 +7,8 @@ import {
 import {
   calculateBiquadEqCoefficients,
   calculateBiquadEqResponse,
-  type MultibandEqBand,
 } from "../../lib/dsp/biquad-eq";
+import { type MultibandEqBand } from "../../lib/dsp/multiband-eq";
 import { clamp, dbToGain, gainToDb } from "../../lib/music";
 import { EQ_CONTROL_LIMITS } from "./eq-control-limits";
